+++ docs/source/conf.py
@@ -13,15 +13,15 @@
 import os
 import sys
 
-sys.path.insert(0, os.path.abspath('.'))
-sys.path.insert(0, os.path.abspath('../..'))
+sys.path.insert(0, os.path.abspath("."))
+sys.path.insert(0, os.path.abspath("../.."))
 
 
 # -- Project information -----------------------------------------------------
 
-project = 'nkululeko'
-copyright = '2023, Felix Burkhardt, Bagus Tris Atmaja'
-author = 'Felix Burkhardt, Bagus Tris Atmaja'
+project = "nkululeko"
+copyright = "2023, Felix Burkhardt, Bagus Tris Atmaja"
+author = "Felix Burkhardt, Bagus Tris Atmaja"
 
 
 # -- General configuration ---------------------------------------------------
@@ -30,16 +30,16 @@ author = 'Felix Burkhardt, Bagus Tris Atmaja'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.viewcode',
+    "sphinx.ext.autodoc",
+    "sphinx.ext.viewcode",
     # 'nbsphinx',           # to embedd ipynb files
-    'sphinx.ext.mathjax',  # to enable latex
-    'sphinx_rtd_theme',
-    'myst_parser',        # to enable markdown
+    "sphinx.ext.mathjax",  # to enable latex
+    "sphinx_rtd_theme",
+    "myst_parser",  # to enable markdown
 ]
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
@@ -52,7 +52,7 @@ exclude_patterns = []
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'sphinx_rtd_theme'
+html_theme = "sphinx_rtd_theme"
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
@@ -61,16 +61,16 @@ html_theme = 'sphinx_rtd_theme'
 
 # suffix for source files
 source_suffix = {
-    '.rst': 'restructuredtext',
-    '.txt': 'restructuredtext',
-    '.md': 'markdown',
+    ".rst": "restructuredtext",
+    ".txt": "restructuredtext",
+    ".md": "markdown",
 }
 
 # headed anchors until h3
 myst_heading_anchors = 3
 
 # The master toctree document.
-master_doc = 'index'
+master_doc = "index"
 
 
 # html_theme_options = {