+++ nkululeko/constants.py
@@ -1,4 +1,4 @@
-VERSION = "1.6.3"
+VERSION = "1.7.0"
 SAMPLING_RATE = 16000
 COL_SEX = "gender"
 COL_AGE = "age"